+++ src/router.php
@@ -299,7 +299,8 @@ function fst_run() {
         $req_header = fst_config('spa.header_request', 'X-FST-Request');
         $target_header = fst_config('spa.header_target', 'X-FST-Target');
         $indicator_class = fst_config('spa.indicator_class', 'fst-loading');
-        $inject_id = $script_id ? 'id="'.$script_id.'" data-req-header="'.$req_header.'" data-target-header="'.$target_header.'" data-indicator-class="'.$indicator_class.'"' : '';
+        $history_cache = fst_config('spa.history_cache', false) ? ' data-history-cache="true"' : '';
+        $inject_id = $script_id ? 'id="'.$script_id.'" data-req-header="'.$req_header.'" data-target-header="'.$target_header.'" data-indicator-class="'.$indicator_class.'"'.$history_cache : '';
         $script_tag = "<script {$inject_id}>\n" . (defined('FST_SPA_JS_CODE') ? FST_SPA_JS_CODE : '') . "\n</script>";
         
         if (stripos($output, '</body>') !== false) {